functions/sow-generation-func/requirements.txt
functions-framework==3.*
# Let pip find the correct, compatible versions for Google libraries
google-cloud-firestore
google-cloud-storage
google-cloud-aiplatform
google-auth